+++ lib/models/display_event.dart
@@ -1,0 +1,12 @@
+enum DisplayEventAction { added, removed, initial }
+
+class DisplayEvent {
+  final DisplayEventAction action;
+  final int publicDisplayCount;
+  final String message;
+
+  DisplayEvent(
+      {required this.action,
+      required this.publicDisplayCount,
+      required this.message});
+}