+++ CHANGELOG.md
@@ -2,3 +2,7 @@
 * Initial release. All planned features working as expected
 ## 1.0.1
 * Flutter formatting for extra pub score
+## 1.3.0
+* Added display change monitoring
+* Added phone call monitoring
+* Added microphone usage monitoring